Move TestingSplit from test to main
