+++ presto-main/src/test/java/com/facebook/presto/execution/TaskTestUtils.java
@@ -38,6 +38,7 @@ import com.facebook.presto.sql.planner.TestingTableHandle;
 import com.facebook.presto.sql.planner.plan.PlanFragmentId;
 import com.facebook.presto.sql.planner.plan.PlanNodeId;
 import com.facebook.presto.sql.planner.plan.TableScanNode;
+import com.facebook.presto.testing.TestingSplit;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
