+++ presto-main/src/test/java/com/facebook/presto/execution/TestingPageSourceProvider.java
@@ -20,6 +20,7 @@ import com.facebook.presto.spi.ConnectorSession;
 import com.facebook.presto.spi.ConnectorSplit;
 import com.facebook.presto.spi.FixedPageSource;
 import com.facebook.presto.spi.Page;
+import com.facebook.presto.testing.TestingSplit;
 import com.google.common.collect.ImmutableList;
 
 import java.util.List;
